move some xcorr tests from test_util to test_cross_correlation
